fixed memory leaks + added tools lib

+++ include/lib/matrix.h
@@ -1,2 +1,2 @@
-void transpose_sequential(int n, double **A, double **B);
-void transpose_parallel(int n, double **A, double **B);
+void transpose_sequential(int n, int **A, int **B);
+void transpose_parallel(int n, int **A, int **B);
