+++ src/matrix.cpp
@@ -1,7 +1,7 @@
 #include <lib/matrix.h>
 #include <omp.h>
 
-void transpose_sequential(int n, double **A, double **B) {
+void transpose_sequential(int n, int **A, int **B) {
   for (int i = 0; i < n; i++) {
     for (int j = 0; j < n; j++) {
       B[i][j] = A[j][i];
@@ -9,7 +9,7 @@ void transpose_sequential(int n, double **A, double **B) {
   }
 }
 
-void transpose_parallel(int n, double **A, double **B) {
+void transpose_parallel(int n, int **A, int **B) {
 #pragma omp parallel for
   for (int i = 0; i < n; i++) {
     for (int j = 0; j < n; j++) {
